+++ lib/mod_manager/homebrew_grid/wispgrid_header_row_view.dart
@@ -10,7 +10,7 @@ import 'package:trios/mod_manager/homebrew_grid/wispgrid_group.dart';
 import 'package:trios/thirdparty/dartx/function.dart';
 import 'package:trios/thirdparty/flutter_context_menu/flutter_context_menu.dart';
 import 'package:trios/utils/extensions.dart';
-import 'package:trios/widgets/MultiSplitViewMixin.dart';
+import 'package:trios/widgets/multi_split_mixin_view.dart';
 import 'package:trios/widgets/hoverable_widget.dart';
 import 'package:trios/widgets/moving_tooltip.dart';